requirements.txt
sounddevice
soundfile
openai-whisper